requirements.txt
opencv-python>=4.9.0.80
diffusers>=0.31.0
transformers>=4.49.0,<=4.51.3
tokenizers>=0.20.3
accelerate>=1.1.1
tqdm
imageio
easydict
ftfy
dashscope
numpy>=1.23.5,<2
scipy
librosa
omegaconf
pandas
open-clip-torch
protobuf
sentencepiece
huggingface_hub>=0.25.0
